+++ test/run_all.sh
@@ -17,3 +17,6 @@ pyinstall-cffi/run_test.sh
 
 # Run test against app built with PyInstaller and includes a helper app
 pyinstall-aux-static-exec/run_test.sh
+
+# Verify that target system libs can't be loaded
+no-default-lib/run_test.sh